Support additional layers
For any layers A, B, ... specified via MIRYOKU_EXTRA_LAYERS (up to 8),

  [A] = LAYOUT_miryoku(MIRYOKU_TABLE_A),
  [B] = LAYOUT_miryoku(MIRYOKU_TABLE_B),
  ...

will be added between the base layer and the top miryoku layers.

The tables can either be defined by the user or the tables that miryoku
defines can be used by specifying e.g. 'QWERTY' for a qwerty layer.